adding tests

+++ internal/controller/controller_filer_service.go
@@ -33,7 +33,7 @@ func (r *SeaweedReconciler) createFilerPeerService(m *seaweedv1.Seaweed) *corev1
 		})
 	}
 	if m.Spec.Filer.IAM {
-		iamPort := seaweedv1.FilerIAMPort
+		iamPort := int32(seaweedv1.FilerIAMPort)
 		if m.Spec.IAM != nil && m.Spec.IAM.Port != nil {
 			iamPort = *m.Spec.IAM.Port
 		}
@@ -97,7 +97,7 @@ func (r *SeaweedReconciler) createFilerService(m *seaweedv1.Seaweed) *corev1.Ser
 		})
 	}
 	if m.Spec.Filer.IAM {
-		iamPort := seaweedv1.FilerIAMPort
+		iamPort := int32(seaweedv1.FilerIAMPort)
 		if m.Spec.IAM != nil && m.Spec.IAM.Port != nil {
 			iamPort = *m.Spec.IAM.Port
 		}
